4.21-4.22 Deleting blogs
- Deletion requires authentification
- All tests working

+++ Osa 4/blogilista/controllers/blogs.js
@@ -2,8 +2,9 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
+const blog = require('../models/blog')
 
-blogsRouter.get('/', async (request, response) => {
+blogsRouter.get('/', async (request, response, next) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name:1 })
   response.json(blogs.map(blog => blog.toJSON()))
 })
@@ -48,7 +49,18 @@ blogsRouter.put('/:id', async (request, response, next) => {
   response.json(update.toJSON())
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', async (request, response, next) => {
+  const body = request.body
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const selectedBlog = await Blog.findById(request.params.id)
+  if ( selectedBlog.user.toString() !== user.id.toString() ){
+    return response.status(401).json({ error: 'that blog is not created by you'})
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
